+++ lib/reactotron-react-native-mmkv/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.2.7](https://github.com/infinitered/reactotron/compare/reactotron-react-native-mmkv@0.2.6...reactotron-react-native-mmkv@0.2.7) (2024-08-26)
+
+### [0.2.6](https://github.com/infinitered/reactotron/compare/reactotron-react-native-mmkv@0.2.5...reactotron-react-native-mmkv@0.2.6) (2024-02-28)
+
 ### [0.2.5](https://github.com/infinitered/reactotron/compare/reactotron-react-native-mmkv@0.2.4...reactotron-react-native-mmkv@0.2.5) (2024-02-14)
 
 ## [0.2.4](https://github.com/infinitered/reactotron/compare/reactotron-react-native-mmkv@0.2.3...reactotron-react-native-mmkv@0.2.4) (2024-02-14)